Rename the namespace RFD.OFXTool.Library.Ofx to Entities
