+++ OFX-Tool.Library/Core/Extract/ExtractSTATUS.cs
@@ -1,4 +1,4 @@
-﻿using RFD.OFXTool.Library.Ofx;
+﻿using RFD.OFXTool.Library.Entities;
 using System.Xml;
 
 namespace RFD.OFXTool.Library.Core.Extract
