+++ OFX-Tool.Library/Core/Extract/ExtractSTMTTRNRS.cs
@@ -1,4 +1,4 @@
-﻿using RFD.OFXTool.Library.Ofx.Bank;
+﻿using RFD.OFXTool.Library.Entities.Bank;
 using System.Xml;
 
 namespace RFD.OFXTool.Library.Core.Extract
