--- OFX-Tool.Library/Entities/BankAccount.cs
@@ -1,10 +1,0 @@
-namespace RFD.OFXTool.Library.Entities
-{
-    public class BankAccount
-    {
-        public string Type { get; set; }
-        public string AgencyCode { get; set; }
-        public Bank Bank { get; set; }
-        public string AccountCode { get; set; }
-    }
-}
